Enabled tracking of the full scope of a log message
Created a CurrentScope property in TeskSink and a FullScope property in LogEntry to track the scopes as entries are made.

Created ScopeTests.cs to test this behaviour.
Moved TestScope to a nested class inside TestSink.

Marked the old scope entry as obsolete.

This solution is NOT thread safe and does not support scopes being disposed out of order.